Split binary format handling into a module
